Refactored app structure
