+++ src/components/NavigationBar/index.js
@@ -2,7 +2,7 @@ import { h } from 'preact'
 import { connect } from 'react-redux'
 import classNames from 'classnames'
 import { compose } from '~utils/func'
-import { setNavigationDisabled } from '../../core/store/actions/globals'
+import { setNavigationDisabled } from '../ReduxAppWrapper/store/actions/globals'
 import { withFullScreenState } from '../FullScreen'
 import style from './style.css'
 import {isDesktop} from '~utils/index'
