+++ src/components/ReduxAppWrapper/constants/index.js
