+++ src/components/ReduxAppWrapper/store/actions/index.js
@@ -6,4 +6,4 @@ import { RESET_STORE } from '../../constants'
 const reset = payload => ({ type: RESET_STORE, payload })
 
 export const unboundActions = {...globals, ...captures, reset}
-export const actions = bindActionCreators(unboundActions, store.dispatch)
+export const actions = bindActionCreators(unboundActions)
