+++ src/components/ReduxAppWrapper/store/reducers/captures.js
