+++ src/components/ReduxAppWrapper/store/reducers/globals.js
