--- src/components/StoreWrapper/index.js
@@ -1,23 +1,0 @@
-import { h, Component } from 'preact'
-import { Provider as ReduxProvider } from 'react-redux'
-import reducer from '../../core/store/reducers'
-import { createStore } from 'redux'
-
-class StoreWrapper extends Component {
-  constructor(props) {
-    super(props)
-    this.store = createStore(reducer,
-      window.__REDUX_DEVTOOLS_EXTENSION__ ? window.__REDUX_DEVTOOLS_EXTENSION__() : undefined
-    )
-  }
-  
-  render() {
-    return (
-      <ReduxProvider store={this.store}>
-        {this.props.children}
-      </ReduxProvider>
-    )
-  }
-}
-
-export default StoreWrapper
