--- src/core/index.js
@@ -1,6 +1,0 @@
-import { actions, unboundActions } from './store/actions'
-
-export {
-  actions,
-  unboundActions,
-}
