--- src/core/store/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux'
-import { captures } from './captures'
-import globals from './globals'
-
-const appReducer = combineReducers({
-  captures,
-  globals
-})
-
-export default (state, action) => appReducer(state, action)
